# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ClassificationResult": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "document_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "is_mock": {
            "default": false,
            "title": "Is Mock",
            "type": "boolean"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "reasoning_summary": {
            "title": "Reasoning Summary",
            "type": "string"
          },
          "requires_review": {
            "title": "Requires Review",
            "type": "boolean"
          }
        },
        "required": [
          "document_type",
          "confidence",
          "reasoning_summary",
          "requires_review",
          "provider",
          "model_name",
          "prompt_version"
        ],
        "title": "ClassificationResult",
        "type": "object"
      },
      "DemoLimits": {
        "properties": {
          "documents_per_day": {
            "title": "Documents Per Day",
            "type": "integer"
          },
          "max_pages": {
            "title": "Max Pages",
            "type": "integer"
          },
          "max_upload_mb": {
            "title": "Max Upload Mb",
            "type": "number"
          },
          "questions_per_day": {
            "title": "Questions Per Day",
            "type": "integer"
          },
          "retention_hours": {
            "title": "Retention Hours",
            "type": "integer"
          }
        },
        "required": [
          "documents_per_day",
          "questions_per_day",
          "max_upload_mb",
          "max_pages",
          "retention_hours"
        ],
        "title": "DemoLimits",
        "type": "object"
      },
...
      "DemoSessionResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "ttl_hours": {
            "title": "Ttl Hours",
            "type": "integer"
          }
        },
        "required": [
          "active",
          "created",
          "ttl_hours"
        ],
        "title": "DemoSessionResponse",
        "type": "object"
      },
      "DemoStatusResponse": {
        "properties": {
          "ai_mode": {
            "enum": [
              "live",
              "offline"
            ],
            "title": "Ai Mode",
            "type": "string"
          },
          "documents_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documents Remaining"
          },
          "limits": {
            "$ref": "#/components/schemas/DemoLimits"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "model_provider": {
            "title": "Model Provider",
            "type": "string"
          },
          "offline_reason": {
            "anyOf": [
              {
                "enum": [
                  "budget_reached",
                  "configured_offline"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offline Reason"
          },
          "questions_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Questions Remaining"
          },
          "session_active": {
            "title": "Session Active",
            "type": "boolean"
          }
        },
        "required": [
          "ai_mode",
          "offline_reason",
          "model_provider",
          "model_name",
          "session_active",
          "documents_remaining",
          "questions_remaining",
          "limits"
        ],
        "title": "DemoStatusResponse",
        "type": "object"
      },
...
      "DocumentResponse": {
        "properties": {
          "char_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char Count"
          },
          "classification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClassificationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Workflow Id"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "document_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "has_text_layer": {
            "title": "Has Text Layer",
            "type": "boolean"
          },
          "page_count": {
            "title": "Page Count",
            "type": "integer"
          },
          "processing": {
            "default": false,
            "description": "True while a background processing run is active",
            "title": "Processing",
            "type": "boolean"
          },
          "security_flags": {
            "items": {
              "type": "string"
            },
            "title": "Security Flags",
            "type": "array"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowStatus"
          },
          "text_extraction_method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TextExtractionMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "filename",
          "status",
          "document_type",
          "classification",
          "page_count",
          "size_bytes",
          "sha256",
          "has_text_layer",
          "text_extraction_method",
          "char_count",
          "security_flags",
          "current_workflow_id",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentResponse",
        "type": "object"
      },
      "DocumentType": {
        "enum": [
          "income_statement",
          "balance_sheet",
          "invoice",
          "bank_statement",
          "fund_summary",
          "unknown"
        ],
        "title": "DocumentType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TextExtractionMethod": {
        "enum": [
          "native_pdf",
          "ocr_tesseract",
          "ocr_textract",
          "none"
        ],
        "title": "TextExtractionMethod",
        "type": "string"
      },
      "UploadResponse": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentResponse"
          },
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          }
        },
        "required": [
          "document",
          "duplicate"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowStatus": {
        "enum": [
          "INGESTED",
          "TEXT_EXTRACTED",
          "CLASSIFIED",
          "ENTITIES_EXTRACTED",
          "VALIDATED",
          "INDEXED",
          "READY",
          "NEEDS_REVIEW",
          "REJECTED",
          "FAILED"
        ],
        "title": "WorkflowStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Ingestion, classification, extraction, validation, grounded Q&A with citations, human review, audit and evaluation for financial documents.",
    "title": "Financial Document Intelligence API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/demo/samples/{sample_id}": {
      "post": {
        "description": "Copy an allow-listed sample into the caller's workspace (same path as an upload).",
        "operationId": "load_sample_demo_samples__sample_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "sample_id",
            "required": true,
            "schema": {
              "title": "Sample Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "cookie",
            "name": "docintel_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Docintel Demo"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Load Sample",
        "tags": [
          "demo"
        ]
      }
    },
    "/demo/samples/{sample_id}/file": {
      "get": {
        "description": "The original sample PDF, for side-by-side viewing.",
        "operationId": "sample_file_demo_samples__sample_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "sample_id",
            "required": true,
            "schema": {
              "title": "Sample Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sample File",
        "tags": [
          "demo"
        ]
      }
    },
    "/demo/session": {
      "post": {
        "description": "Start (or keep) an anonymous visitor session bound to a private workspace.",
        "operationId": "start_session_demo_session_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "docintel_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Docintel Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Session",
        "tags": [
          "demo"
        ]
      }
    },
    "/demo/status": {
      "get": {
        "description": "Which AI engine is answering right now, and the visitor's remaining allowance.",
        "operationId": "demo_status_demo_status_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "docintel_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Docintel Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo Status",
        "tags": [
          "demo"
        ]
      }
    },
...
  }
}